views: Move invitation views to their own file.
